Rename module types to typehints
